Implemented Clerk authentication

+++ backend/Controller/taskController.ts
@@ -1,13 +1,22 @@
 import Task from "../Model/taskModel";
+import { getAuth } from "@clerk/express";
 
 export const createTask = async (req: any, res: any) => {
     try {
+        const { userId } = getAuth(req)
+
+        if (!userId) {
+            return res.status(401).json({
+                message: "Unauthorized"
+            });
+        }
+
         const { task, status } = req.body;
-        const newTask = new Task({ 
-            user: req.user.userId,
+        const newTask = new Task({
+            clerkId: userId,
             task: task,
             status: status
-             });
+        });
         await newTask.save();
         res.status(201).json(newTask);
     } catch (error) {
@@ -17,7 +26,15 @@ export const createTask = async (req: any, res: any) => {
 
 export const getAllTasks = async (req: any, res: any) => {
     try {
-        const tasks = await Task.find(req.user ? { user: req.user.userId } : {}).sort({ createdAt: -1 });
+        const { userId } = getAuth(req)
+
+        if (!userId) {
+            return res.status(401).json({
+                message: "Unauthorized"
+            });
+        }
+
+        const tasks = await Task.find({ clerkId: userId }).sort({ createdAt: -1 });
         res.status(200).json(tasks);
     } catch (error) {
         res.status(500).json({ message: "Error fetching tasks", error });
@@ -26,6 +43,14 @@ export const getAllTasks = async (req: any, res: any) => {
 
 export const updateTask = async (req: any, res: any) => {
     try {
+        const { userId } = getAuth(req)
+
+        if (!userId) {
+            return res.status(401).json({
+                message: "Unauthorized"
+            });
+        }
+
         const { id } = req.params;
         const { task, status } = req.body;
         const updatedTask = await Task.findByIdAndUpdate(id, { task, status }, { new: true });
@@ -40,6 +65,14 @@ export const updateTask = async (req: any, res: any) => {
 
 export const deleteTask = async (req: any, res: any) => {
     try {
+        const { userId } = getAuth(req)
+
+        if (!userId) {
+            return res.status(401).json({
+                message: "Unauthorized"
+            });
+        }
+
         const { id } = req.params;
         const deletedTask = await Task.findByIdAndDelete(id);
         if (!deletedTask) {
